S_stroke() size int -> double
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@38646 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ display/d.grid/fiducial.c
@@ -26,7 +26,7 @@ void plot_symbol(double easting, double northing, int color, double rotation,
     SYMBOL *Symb;
     RGBA_Color *line_color, *fill_color;
     int R, G, B;
-    int size = 16;
+    double size = 16.0;
     int tolerance = 0;
 
     line_color = G_malloc(sizeof(RGBA_Color));
